add auth hook
